Improve usability of generate scripts
Before the generate_data and generate_queries scripts would leave
behind broken gzip files if the underlying command failed. This
then required the user to remove the file before trying again,
otherwise the script would just detect the old (broken) file and skip
the step. This commit adds a 'trap' to cleanup any files when an
error occurs in generation.