Split TS load data processing into own function
To improve the testability of the tsbs_load_timescaledb binary and
reduce the complexity of some of its most complex function, we split
the processing of the insertData slice into tags and data into its
own function. This separates a non-DB task from a DB task so that
the former can be tested without an adequate mock for the latter.