Change default config because Viper doesn't use defaults to override
If a config structure is used to populate a Viper instance,
and a flag is bound to a property of that structure, the default
value of the flag does not override the empty property in the struct.
This commit modifies the code that generates the sample config
so it serializes the default values of each flag.

Also, update CI to only test on Go 1.13 and 1.14