Fix tests that were broken with the introduction of DataSource
A lot of the code was moved to new packages and generaly
re-organized. This together with the introduction of the
DataSource and ImplementedTarget as new interfaces that
decouple db-specific code and the loaders required
the tests to be moved to new packages, reduce their dependence
on globaly available variables, and introduce mocks where
target specific serializers were previously used.