Allow prometheus simulated data source to use current time in load
In order to better support continuous loading over larger
periods of times where the generated data should not go
into the future or trail too far into the past, the prometheus
simulated data source can be configured to use the current time
for the generated metrics.
The simulator can be configured with any time range in order
to choose the number of generated points, but the generated timestamps
will not be used.