Prevent simulator from generating duplicate metrics
Because of the ability to use the current timestamp
when simulating prometheus metrics and prometheus using
ms time granularity, the simulator created duplicates with
the same timestamp and series_id.
This commit introduces a mechanism that increments the ts
by at least a milisecond or fast forwards if the current time
is no longer lagging.