Allow disabling space partitioning
TSBS uses space partitioning with only 1 partitioning which is a
configuration that provides no benefits at all. It also doesn't
allow disabling space partitioning which is unfortunate as not all
timescaledb optimization work with space partitioning.

This patch adds support for creating hypertables without space
partitions by setttings `--partitions` to `0`. This patch also
changes the default for partitions to `0`.